+++ src/main/java/com/devcourse/springjpaboard/application/post/controller/dto/CreatePostRequest.java
@@ -1,0 +1,5 @@
+package com.devcourse.springjpaboard.application.post.controller.dto;
+
+public record CreatePostRequest(String title, String content, Long userId) {
+
+}
