+++ src/main/java/com/devcourse/springjpaboard/application/post/controller/dto/UpdatePostRequest.java
@@ -1,0 +1,4 @@
+package com.devcourse.springjpaboard.application.post.controller.dto;
+
+public record UpdatePostRequest(String title, String content) {
+}
