+++ src/main/java/com/devcourse/springjpaboard/application/post/controller/dto/UserDto.java
@@ -1,0 +1,4 @@
+package com.devcourse.springjpaboard.application.post.controller.dto;
+
+public record UserDto(Long id, String name, int age, String hobby) {
+}
