+++ src/main/java/com/devcourse/springjpaboard/application/post/model/Post.java
@@ -1,0 +1,38 @@
+package com.devcourse.springjpaboard.application.post.model;
+
+import com.devcourse.springjpaboard.core.model.BaseEntity;
+import com.devcourse.springjpaboard.application.user.model.User;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.util.Objects;
+
+@Entity
+@Table(name = "post")
+@Getter
+@Setter
+public class Post extends BaseEntity {
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @Column(nullable = false)
+    private String title;
+
+    @Lob
+    @Column(nullable = false)
+    private String content;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    public void setUser(User user) {
+        if (Objects.nonNull(this.user)) {
+            this.user.getPosts().remove(this);
+        }
+        this.user = user;
+        user.getPosts().add(this);
+    }
+}
