+++ src/main/java/com/devcourse/springjpaboard/application/post/service/dto/PostResponse.java
@@ -1,0 +1,4 @@
+package com.devcourse.springjpaboard.application.post.service.dto;
+
+public record PostResponse(String title, String content) {
+}
