+++ src/main/java/com/devcourse/springjpaboard/application/user/controller/dto/CreateUserRequest.java
@@ -1,0 +1,4 @@
+package com.devcourse.springjpaboard.application.user.controller.dto;
+
+public record CreateUserRequest(String name, int age, String hobby) {
+}
