+++ src/main/java/com/devcourse/springjpaboard/application/user/controller/dto/UserResponse.java
@@ -1,0 +1,4 @@
+package com.devcourse.springjpaboard.application.user.controller.dto;
+
+public record UserResponse(Long id, String name, int age, String hobby) {
+}
