+++ src/main/java/com/devcourse/springjpaboard/application/user/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package com.devcourse.springjpaboard.application.user.repository;
+
+import com.devcourse.springjpaboard.application.user.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
