+++ src/main/java/com/devcourse/springjpaboard/core/exception/ExceptionMessage.java
@@ -1,0 +1,10 @@
+package com.devcourse.springjpaboard.core.exception;
+
+public class ExceptionMessage {
+
+    private ExceptionMessage() {}
+
+    public static String NOT_FOUND_POST = "해당하는 게시글을 찾을 수 없습니다.";
+
+    public static String NOT_FOUND_USER = "해당하는 유저를 찾을 수 없습니다.";
+}
