+++ src/main/java/com/devcourse/springjpaboard/core/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package com.devcourse.springjpaboard.core.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
