--- src/main/java/com/devcourse/springjpaboard/exception/ExceptionMessage.java
@@ -1,10 +1,0 @@
-package com.devcourse.springjpaboard.exception;
-
-public class ExceptionMessage {
-
-    private ExceptionMessage() {}
-
-    public static String NOT_FOUND_POST = "해당하는 게시글을 찾을 수 없습니다.";
-
-    public static String NOT_FOUND_USER = "해당하는 유저를 찾을 수 없습니다.";
-}
