--- src/main/java/com/devcourse/springjpaboard/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package com.devcourse.springjpaboard.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
