--- src/main/java/com/devcourse/springjpaboard/model/BaseEntity.java
@@ -1,34 +1,0 @@
-package com.devcourse.springjpaboard.model;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@MappedSuperclass
-public class BaseEntity {
-
-    @CreatedDate
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-
-    @CreatedBy
-    @Column(name = "created_by", nullable = false)
-    private String createdBy;
-
-    @LastModifiedBy
-    @Column(name = "last_modified_by", nullable = false)
-    private String lastModifiedBy;
-
-    @LastModifiedDate
-    @Column(name = "last_modified_data", nullable = false)
-    private LocalDateTime lastModifiedData;
-}
