--- src/main/java/com/devcourse/springjpaboard/model/post/Post.java
@@ -1,38 +1,0 @@
-package com.devcourse.springjpaboard.model.post;
-
-import com.devcourse.springjpaboard.model.BaseEntity;
-import com.devcourse.springjpaboard.model.user.User;
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.util.Objects;
-
-@Entity
-@Table(name = "post")
-@Getter
-@Setter
-public class Post extends BaseEntity {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Lob
-    @Column(nullable = false)
-    private String content;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    public void setUser(User user) {
-        if (Objects.nonNull(this.user)) {
-            this.user.getPosts().remove(this);
-        }
-        this.user = user;
-        user.getPosts().add(this);
-    }
-}
