--- src/main/java/com/devcourse/springjpaboard/post/controller/dto/CreatePostRequest.java
@@ -1,5 +1,0 @@
-package com.devcourse.springjpaboard.post.controller.dto;
-
-public record CreatePostRequest(String title, String content, Long userId) {
-
-}
