--- src/main/java/com/devcourse/springjpaboard/post/controller/dto/UpdatePostRequest.java
@@ -1,4 +1,0 @@
-package com.devcourse.springjpaboard.post.controller.dto;
-
-public record UpdatePostRequest(String title, String content) {
-}
