--- src/main/java/com/devcourse/springjpaboard/post/controller/dto/UserDto.java
@@ -1,4 +1,0 @@
-package com.devcourse.springjpaboard.post.controller.dto;
-
-public record UserDto(Long id, String name, int age, String hobby) {
-}
