--- src/main/java/com/devcourse/springjpaboard/post/service/dto/PostResponse.java
@@ -1,6 +1,0 @@
-package com.devcourse.springjpaboard.post.service.dto;
-
-import com.devcourse.springjpaboard.post.controller.dto.UserDto;
-
-public record PostResponse(String title, String content) {
-}
