--- src/main/java/com/devcourse/springjpaboard/user/controller/dto/CreateUserRequest.java
@@ -1,4 +1,0 @@
-package com.devcourse.springjpaboard.user.controller.dto;
-
-public record CreateUserRequest(String name, int age, String hobby) {
-}
