--- src/main/java/com/devcourse/springjpaboard/user/controller/dto/UserResponse.java
@@ -1,4 +1,0 @@
-package com.devcourse.springjpaboard.user.controller.dto;
-
-public record UserResponse(Long id, String name, int age, String hobby) {
-}
