+++ src/test/java/com/devcourse/springjpaboard/post/service/PostServiceTest.java
@@ -1,14 +1,18 @@
 package com.devcourse.springjpaboard.post.service;
 
-import com.devcourse.springjpaboard.model.post.Post;
-import com.devcourse.springjpaboard.model.user.User;
-import com.devcourse.springjpaboard.post.controller.dto.CreatePostRequest;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doReturn;
+
+import com.devcourse.springjpaboard.application.post.model.Post;
+import com.devcourse.springjpaboard.application.post.service.PostService;
+import com.devcourse.springjpaboard.application.user.model.User;
+import com.devcourse.springjpaboard.application.post.controller.dto.CreatePostRequest;
 import com.devcourse.springjpaboard.post.controller.stub.PostStubs;
-import com.devcourse.springjpaboard.post.converter.PostConverter;
-import com.devcourse.springjpaboard.post.repository.PostRepository;
-import com.devcourse.springjpaboard.post.service.dto.PostResponse;
+import com.devcourse.springjpaboard.application.post.converter.PostConverter;
+import com.devcourse.springjpaboard.application.post.repository.PostRepository;
+import com.devcourse.springjpaboard.application.post.service.dto.PostResponse;
 import com.devcourse.springjpaboard.user.controller.stub.UserStubs;
-import com.devcourse.springjpaboard.user.repository.UserRepository;
+import com.devcourse.springjpaboard.application.user.repository.UserRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
@@ -21,81 +25,78 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doReturn;
-
 
 @ExtendWith(MockitoExtension.class)
 @Disabled
 class PostServiceTest {
 
-    private final ObjectMapper mapper = new ObjectMapper();
+  private final ObjectMapper mapper = new ObjectMapper();
 
-    private MockMvc mockMvc;
+  private MockMvc mockMvc;
 
-    @Mock
-    private PostRepository postRepository;
+  @Mock
+  private PostRepository postRepository;
 
-    @Mock
-    private UserRepository userRepository;
+  @Mock
+  private UserRepository userRepository;
 
-    @Mock
-    private PostConverter postConverter;
+  @Mock
+  private PostConverter postConverter;
 
-    @InjectMocks
-    private PostService postService;
+  @InjectMocks
+  private PostService postService;
 
-    @BeforeEach
-    void init() {
-        mockMvc = MockMvcBuilders.standaloneSetup(postService).build();
-    }
+  @BeforeEach
+  void init() {
+    mockMvc = MockMvcBuilders.standaloneSetup(postService).build();
+  }
 
-    @Test
-    @DisplayName("게시글 저장 테스트")
-    void saveTest() {
-        // given
-        CreatePostRequest request = PostStubs.createPostRequest();
-        User responseUser = UserStubs.findUser(request.userId());
-        Post responsePost = PostStubs.post(responseUser);
-        PostResponse response = PostStubs.createPostResponse();
-        doReturn(responseUser)
-                .when(userRepository)
-                .findById(any(Long.class));
-        doReturn(responsePost)
-                .when(postConverter)
-                .convertPostRequest(any(CreatePostRequest.class), any(User.class));
+  @Test
+  @DisplayName("게시글 저장 테스트")
+  void saveTest() {
+    // given
+    CreatePostRequest request = PostStubs.createPostRequest();
+    User responseUser = UserStubs.findUser(request.userId());
+    Post responsePost = PostStubs.post(responseUser);
+    PostResponse response = PostStubs.createPostResponse();
 
-        // when
-        // then
-    }
+    // when
+    doReturn(responseUser)
+        .when(userRepository)
+        .findById(any(Long.class));
+    doReturn(responsePost)
+        .when(postConverter)
+        .convertPostRequest(any(CreatePostRequest.class), any(User.class));
+    // then
+  }
 
-    @Test
-    @DisplayName("게시글 번호로 게시글을 찾을 수 있는지 테스트")
-    void findOneTest() {
-        // given
+  @Test
+  @DisplayName("게시글 번호로 게시글을 찾을 수 있는지 테스트")
+  void findOneTest() {
+    // given
 
-        // when
+    // when
 
-        // then
-    }
+    // then
+  }
 
-    @Test
-    @DisplayName("지정한 페이지만큼 게시글을 찾는지 테스트")
-    void findAllTest() {
-        // given
+  @Test
+  @DisplayName("지정한 페이지만큼 게시글을 찾는지 테스트")
+  void findAllTest() {
+    // given
 
-        // when
+    // when
 
-        // then
-    }
+    // then
+  }
 
-    @Test
-    @DisplayName("게시글 업데이트가 정상적으로 되는지 테스트")
-    void updateTest() {
-        // given
+  @Test
+  @DisplayName("게시글 업데이트가 정상적으로 되는지 테스트")
+  void updateTest() {
+    // given
 
-        // when
+    // when
 
-        // then
-    }
+    // then
+  }
 }
